Handel 1.10 version

+++ dxpApp/handelSrc/handel_dyn_detector.c
@@ -1,15 +1,6 @@
 /*
- * handel_dyn_detector.c
- *
- *
- * Created 10/03/01 -- PJF
- *
- * This file is nothing more then the routines
- * that were once in the (now non-existent) file
- * handel_dynamic_config.c.
- *
- * Copyright (c) 2002,2003,2004, X-ray Instrumentation Associates
- *               2005, XIA LLC
+ * Copyright (c) 2002-2004, X-ray Instrumentation Associates
+ *               2005-2010, XIA LLC
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, 
@@ -42,6 +33,8 @@
  * THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF 
  * SUCH DAMAGE.
  *
+ * $Id: handel_dyn_detector.c 15170 2010-03-27 01:26:34Z patrick $
+ *
  */
 
 
@@ -53,13 +46,13 @@
 #include "xerxes.h"
 #include "xerxes_errors.h"
 #include "xerxes_structures.h"
-
 #include "xia_handel.h"
+#include "xia_handel_structures.h"
+
 #include "handel_generic.h"
 #include "handeldef.h"
 #include "handel_errors.h"
-#include "xia_handel_structures.h"
-
+#include "handel_log.h"
 
 
 /** @brief Creates a new detector
@@ -224,7 +217,7 @@ HANDEL_EXPORT int HANDEL_API xiaAddDetectorItem(char *alias, char *name,
 /* convert the name to lower case */
 	for (i = 0; i < (unsigned int)strlen(name); i++)
 	{
-		strtemp[i] = (char)tolower(name[i]);
+		strtemp[i] = (char)tolower((int)name[i]);
 	}
 	strtemp[strlen(name)] = '\0';
 
@@ -421,7 +414,7 @@ HANDEL_EXPORT int HANDEL_API xiaModifyDetectorItem(char *alias, char *name, void
 /* convert the name to lower case */
 	for (i = 0; i < (unsigned int)strlen(name); i++)
   	{
-		strtemp[i] = (char)tolower(name[i]);
+		strtemp[i] = (char)tolower((int)name[i]);
 	}
 	strtemp[strlen(name)] = '\0';
 
@@ -518,7 +511,7 @@ HANDEL_EXPORT int HANDEL_API xiaGetDetectorItem(char *alias, char *name, void
 /* Convert name to lowercase */
 	for (i = 0; i < (unsigned int)strlen(name); i++)
 	{
-		strtemp[i] = (char)tolower(name[i]);
+		strtemp[i] = (char)tolower((int)name[i]);
 	}
 	strtemp[strlen(name)] = '\0';
 
@@ -728,7 +721,7 @@ HANDEL_EXPORT int HANDEL_API xiaRemoveDetector(char *alias)
 /* Turn the alias into lower case version, and terminate with a null char */
 	for (i = 0; i < (int)strlen(alias); i++)
 	{
-		strtemp[i] = (char)tolower(alias[i]);
+		strtemp[i] = (char)tolower((int)alias[i]);
 	}
 	strtemp[strlen(alias)] = '\0';
 
@@ -788,7 +781,7 @@ HANDEL_SHARED Detector* HANDEL_API xiaFindDetector(char *alias)
 /* Turn the alias into lower case version, and terminate with a null char */
 	for (i = 0; i < (int)strlen(alias); i++)
 	{
-		strtemp[i] = (char)tolower(alias[i]);
+		strtemp[i] = (char)tolower((int)alias[i]);
 	}
 	strtemp[strlen(alias)] = '\0';
